+++ internal/topic/topicwriterinternal/writer_single_stream_test.go
@@ -1,6 +1,7 @@
 package topicwriterinternal
 
 import (
+	"encoding/json"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -39,5 +40,62 @@ func TestWriterImpl_CreateInitMessage(t *testing.T) {
 			SingleStreamWriterConfig{getLastSeqNum: false},
 		)
 		require.False(t, w.createInitRequest().GetLastSeqNo)
+	})
+}
+
+func TestCutRequestBytes(t *testing.T) {
+	t.Run("SingleLargeMessage", func(t *testing.T) {
+		orig := &rawtopicwriter.WriteRequest{Messages: []rawtopicwriter.MessageData{
+			{Data: make([]byte, 1000)},
+		}}
+		head, rest := cutRequestBytes(orig, 10)
+		require.Same(t, orig, head)
+		require.Nil(t, rest)
+	})
+	t.Run("SingleSmallMessage", func(t *testing.T) {
+		orig := &rawtopicwriter.WriteRequest{Messages: []rawtopicwriter.MessageData{
+			{Data: make([]byte, 1000)},
+		}}
+		head, rest := cutRequestBytes(orig, 1000)
+		require.Same(t, orig, head)
+		require.Nil(t, rest)
+	})
+	t.Run("TwoSmall", func(t *testing.T) {
+		orig := &rawtopicwriter.WriteRequest{Messages: []rawtopicwriter.MessageData{
+			{Data: make([]byte, 100)},
+			{Data: make([]byte, 200)},
+		}}
+		head, rest := cutRequestBytes(orig, 1000)
+		require.Same(t, orig, head)
+		require.Nil(t, rest)
+	})
+	t.Run("Split", func(t *testing.T) {
+		orig := &rawtopicwriter.WriteRequest{Messages: []rawtopicwriter.MessageData{
+			{Data: make([]byte, 100)},
+			{Data: make([]byte, 200)},
+			{Data: make([]byte, 300)},
+		}}
+
+		head, rest := cutRequestBytes(orig, 300)
+
+		JSON := func(a any) string {
+			res, _ := json.MarshalIndent(a, "", "  ")
+
+			return string(res)
+		}
+
+		// it will take only one message, becase it has overhead for grpc messages struct
+		expectedHead := JSON(&rawtopicwriter.WriteRequest{Messages: []rawtopicwriter.MessageData{
+			{Data: make([]byte, 100)},
+		}})
+		expectedRest := JSON(&rawtopicwriter.WriteRequest{Messages: []rawtopicwriter.MessageData{
+			{Data: make([]byte, 200)},
+			{Data: make([]byte, 300)},
+		}})
+
+		headJSON := JSON(head)
+		restJSON := JSON(rest)
+		require.Equal(t, expectedHead, headJSON)
+		require.Equal(t, expectedRest, restJSON)
 	})
 }